Add function comm_readFully() in comm.c
When a process sends a big message to another process which is doing
a comm_read() on this connection, comm_read() returns after reading
around 1500 bytes when both processes are on different machines and
around 16000 bytes when they are on the same machine. Thus we create
comm_readFully() which waits till it has read the required length.
We adapt tests/integration/clientServer/client.c and
tests/integration/clientServer/server.c to test this evolution.
NB: To be perfect, we should also adpat the unit test