Use gccmakedep instead of makedepend
makedepend doesn't find stddef.h because it searches in the wrong directories.
The bug seems pretty old but is not properly fixed.
Moreover, gccmakdep is just the same as "GCC -M", so it doesn't need any
other additional library than gcc to be installed to work, which is not
the case of makedepend (part of xutils-dev).