Add a GCCMAKEDEP variable in the Makefile
For Mac OX S users: the default compiler is llvm-gcc, not gcc.
/usr/bin/gcc is a symbolic link pointing to llvm-gcc-xx.
So the compiler will complain when it tries to use gccmakedep and
the -pthread flag.
A quick fix can be to
1) change the target of the symbolic link towards a installed version of
gcc (preferably not the version of gcc provided by Apple)
2) give the path of a version of gccmakedep e.g. installed via macports
like this: make GCCMAKEDEP=/my/path/for/gccmakedep
OR do export GCCMAKDEPEP=/my/path/for/gccmakdep before executing the
Makefile (recommended for successive builds)

For the other users not concerned by this issue, make will take the
default compiler and gccmakedep binary.