Replacing sem_init() by sem_open()
On Mac OS X, named semaphore have to be used instead of unnamed
semaphores. Indeed, while <semaphore.h> declares sem_init() so that it
compiles properly on OS X, it returns -1 with errno set to ENOSYS
(function not implemented).
So we have to use sem_open() instead of sem_init(), and sem_close() and
sem_unlink() instead of sem_destroy().