JNI perf test
create tests/intergration/JNIPerf providing a JNI interface
to get access to the variable "counters" allowing to get
some data about the performance of the trains protocol from
the C process running the trains protocol to the C process
analysing the data.
Basically, we pass the "counters" structure to JNI
as a byteArray in "dumpDataCounters" in
interface.c. It is stored as a black box in the object attribute data of
the class Counters. And then we get it back in the other
processus running JNIPerf/JNIPerf.c