Suppress compilation warnings on 32 bits OS
- When compiling on a 32 bits OS, there were warnings related to printf
of sizeof. These warnings were due to the fact that sizeof returns a
long unsigned in 64 bits and a plain unsigned in 32 bits. Thus
we used %lu not to have any warnings when compiling on a 64 bits OS.
But %lu generates a warning on 32 bits OS.
To correct this problem, we replace %lu by %zu, as %zu is a format designed
to avoid compilation warnings related to sizeof different size on 32 bits
and 64 bits OS.